--- src/app/themes/page.tsx
@@ -1,11 +1,0 @@
-import ThemesLayout from '@/features/themes/components/ThemesLayout'
-
-function Themes() {
-  return (
-    <div>
-      <ThemesLayout />
-    </div>
-  )
-}
-
-export default Themes
