+++ src/app/timeline/page.tsx
@@ -2,10 +2,7 @@ import TimelineLayout from '@/features/pictures/components/TimelineLayout'
 
 function Timeline() {
   return (
-    <div>
-      <div className='py-5 text-center'>
-        <p className='p-2'>タイムライン</p>
-      </div>
+    <div className='mt-5'>
       <TimelineLayout />
     </div>
   )
